Split TH-D7, TH-D7G, and TH-D700 into a separate file
This is in preparation of a settings overhaul.
